+++ frontend/src/views/AdminVendors.vue
@@ -5,7 +5,7 @@
     </h1>
     <div class="">
       <v-btn
-        class="bg-secondary text-primary me-1 "
+        class="bg-primary me-1 "
         icon
         size="small"
         border="primary thin"
@@ -29,7 +29,7 @@
         </svg>
       </v-btn>
       <v-btn
-        class="bg-secondary text-primary me-1 "
+        class="bg-primary me-1 "
         icon
         size="small"
         border="primary thin"
@@ -81,7 +81,12 @@
               {{ vendor.name }}
             </td>
             <td>
-              {{ vendor.active }}
+              <v-checkbox-btn
+                v-model="vendor.active"
+                color="primary"
+                readonly
+                disabled
+              />
             </td>
             <td>
               <div
@@ -127,6 +132,15 @@
                 >
                   <path d="M9.405 1.05c-.413-1.4-2.397-1.4-2.81 0l-.1.34a1.464 1.464 0 0 1-2.105.872l-.31-.17c-1.283-.698-2.686.705-1.987 1.987l.169.311c.446.82.023 1.841-.872 2.105l-.34.1c-1.4.413-1.4 2.397 0 2.81l.34.1a1.464 1.464 0 0 1 .872 2.105l-.17.31c-.698 1.283.705 2.686 1.987 1.987l.311-.169a1.464 1.464 0 0 1 2.105.872l.1.34c.413 1.4 2.397 1.4 2.81 0l.1-.34a1.464 1.464 0 0 1 2.105-.872l.31.17c1.283.698 2.686-.705 1.987-1.987l-.169-.311a1.464 1.464 0 0 1 .872-2.105l.34-.1c1.4-.413 1.4-2.397 0-2.81l-.34-.1a1.464 1.464 0 0 1-.872-2.105l.17-.31c.698-1.283-.705-2.686-1.987-1.987l-.311.169a1.464 1.464 0 0 1-2.105-.872zM8 10.93a2.929 2.929 0 1 1 0-5.86 2.929 2.929 0 0 1 0 5.858z" />
                 </svg>
+              </div>
+              <div
+                class="btn text-primary"
+                title="Menük"
+                @click="openVendorMenuManager(vendor.id)"
+              >
+                <v-icon>
+                  mdi-food
+                </v-icon>
               </div>
             </td>
           </tr>
@@ -181,6 +195,9 @@ export default {
       },
       openVendorConfiguration: function (id) {
         this.$router.push({ path:`/admin/${id}/config`})
+      },
+      openVendorMenuManager: function (id) {
+        this.$router.push({ path:`/admin/${id}/menu`})
       },
       addVendor: function () {
         this.$router.push({ path:`/admin/add`})
