+++ frontend/src/views/VendorSettings.vue
